rearrange sources to fix test failures
test.py was failing in 0.5 because it was run from the source tree's root
directory, and it found the source tree's ed25519/ directory there (which
does not contain the generated _ed25519.so).

There is now a src/ directory with three subdirectories: ed25519/ (which
contains the python code), ed25519-glue/ (which contains the ed25519module.c
python-to-C glue code), and ed25519-supercop-ref/ (which contains the
original C implementation of Ed25519). The tests were moved into
src/ed25519/test.py .

The large Known-Answer-Tests (which include about 2.5MB of known answers) are
not held in src/ , since I want to keep the installed size as small as
possible. They can be run from a build source tree, but not from an
installation. The normal unit tests *can* be run from an installation, by
just executing test.py as a script.

All imports are now relative (e.g. "from . import _ed25519").